change trunk_id type from uint32_t to uint64_t

+++ src/binlog/trunk/space_log_reader.c
@@ -142,7 +142,7 @@ static int parse_to_skiplist(DASpaceLogReader *reader,
 }
 
 int da_space_log_reader_load(DASpaceLogReader *reader,
-        const uint32_t trunk_id, UniqSkiplist **skiplist)
+        const uint64_t trunk_id, UniqSkiplist **skiplist)
 {
     int result;
     int fd;
