+++ src/trunk/trunk_hashtable.c
@@ -145,7 +145,7 @@ int da_trunk_hashtable_add(DATrunkHTableContext *ctx,
 }
 
 DATrunkFileInfo *da_trunk_hashtable_get_ex(DATrunkHTableContext *ctx,
-        const uint32_t trunk_id, const int log_level)
+        const uint64_t trunk_id, const int log_level)
 {
     DATrunkFileInfo *current;
     int result;
@@ -175,7 +175,7 @@ DATrunkFileInfo *da_trunk_hashtable_get_ex(DATrunkHTableContext *ctx,
             da_ctx = fc_list_entry(ctx, DAContext, trunk_htable_ctx);
             log_it_ex(&g_log_context, log_level,
                     "file: "__FILE__", line: %d, %s "
-                    "trunk id: %u NOT exist", __LINE__,
+                    "trunk id: %"PRId64" NOT exist", __LINE__,
                     da_ctx->module_name, trunk_id);
         }
         return NULL;
